requirements.txt
streamlit
python-dotenv
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
PyPDF2
faiss-cpu


